Simplify ledger
