+++ src/core/providers/ProviderFactory.ts
@@ -7,7 +7,6 @@ import { setAccountProvider } from './accountProvider';
 import { DappProvider } from './DappProvider/DappProvider';
 import { createCrossWindowProvider } from './helpers/crossWindow/createCrossWindowProvider';
 import { createExtensionProvider } from './helpers/extension/createExtensionProvider';
-import { getConfig } from './helpers/getConfig';
 import { createIframeProvider } from './helpers/iframe/createIframeProvider';
 import { createLedgerProvider } from './helpers/ledger/createLedgerProvider';
 import {
@@ -25,12 +24,10 @@ export class ProviderFactory {
   }
 
   public static async create({
-    type,
-    config: userConfig
+    type
   }: IProviderFactory): Promise<DappProvider> {
     let createdProvider: IProvider | null = null;
-    const config = await getConfig(userConfig);
-    const { account, UI } = config;
+    const address = getAddress();
 
     switch (type) {
       case ProviderTypeEnum.extension: {
@@ -44,7 +41,7 @@ export class ProviderFactory {
 
       case ProviderTypeEnum.crossWindow: {
         const provider = await createCrossWindowProvider({
-          address: account?.address
+          address
         });
         createdProvider = provider as unknown as IProvider;
 
@@ -54,7 +51,7 @@ export class ProviderFactory {
       }
 
       case ProviderTypeEnum.ledger: {
-        const ledgerProvider = await createLedgerProvider(UI.ledger.mount);
+        const ledgerProvider = await createLedgerProvider();
 
         if (!ledgerProvider) {
           throw new Error('Unable to create ledger provider');
@@ -78,7 +75,7 @@ export class ProviderFactory {
 
       case ProviderTypeEnum.metamask: {
         const provider = await createIframeProvider({
-          address: account?.address,
+          address,
           type: IframeLoginTypes.metamask
         });
 
@@ -95,7 +92,7 @@ export class ProviderFactory {
 
       case ProviderTypeEnum.passkey: {
         const provider = await createIframeProvider({
-          address: account?.address,
+          address,
           type: IframeLoginTypes.passkey
         });
 
@@ -113,7 +110,7 @@ export class ProviderFactory {
       default: {
         for (const customProvider of this._customProviders) {
           if (customProvider.type === type) {
-            createdProvider = await customProvider.constructor(config);
+            createdProvider = await customProvider.constructor();
             createdProvider.getType = () => type;
           }
         }
