+++ src/core/providers/helpers/restoreProvider.ts
@@ -1,27 +1,17 @@
-import { getAddress } from 'core/methods/account/getAddress';
 import { providerTypeSelector } from 'store/selectors';
 import { getState } from 'store/store';
 import { setAccountProvider } from '../accountProvider';
 import { ProviderFactory } from '../ProviderFactory';
-import { IProviderConfig } from '../types/providerFactory.types';
 
 export async function restoreProvider() {
   const type = providerTypeSelector(getState());
-  const address = getAddress();
 
   if (!type) {
     return;
   }
 
-  const config: IProviderConfig = {
-    account: {
-      address
-    }
-  };
-
   const provider = await ProviderFactory.create({
-    type,
-    config
+    type
   });
 
   if (!provider) {
