+++ src/lib/sdkDappCoreUi.ts
@@ -1,0 +1,14 @@
+import type { CustomElementsDefineOptions } from '@multiversx/sdk-dapp-core-ui/loader';
+export type { LedgerConnectModal } from '@multiversx/sdk-dapp-core-ui/dist/components/ledger-connect-modal';
+
+export async function defineCustomElements(
+  win?: Window,
+  opts?: CustomElementsDefineOptions
+): Promise<void> {
+  try {
+    const loader = await import('@multiversx/sdk-dapp-core-ui/loader');
+    loader.defineCustomElements(win, opts);
+  } catch (err) {
+    throw new Error('@multiversx/sdk-dapp-core-ui not found' + err);
+  }
+}
